+++ spec/support_table_cache_spec.rb
@@ -54,11 +54,34 @@ describe SupportTableCache do
       expect(TestModel.find_by(group: "First", code: "one").value).to eq 1
     end
 
+    it "uses the cache when using find_by!" do
+      expect(TestModel.find_by!(name: "One")).to eq record_1
+      expect(SupportTableCache.cache.read(SupportTableCache.cache_key(TestModel, {name: "One"}, ["name"], true))).to eq record_1
+    end
+
+    it "raises and error when using find_by! and the record doesn't exist" do
+      expect { TestModel.find_by!(name: "Not Exist") }.to raise_error(ActiveRecord::RecordNotFound)
+    end
+
     it "does not use the cache when finding by a single attribute in a composite key" do
       expect(TestModel.find_by(code: "one")).to eq record_1
       expect(SupportTableCache.cache.read(SupportTableCache.cache_key(TestModel, {code: "one"}, ["code"], false))).to eq nil
     end
 
+    it "does not use the cache when finding by a non-cacheable attribute" do
+      expect(SupportTableCache.cache).to receive(:fetch).and_return(:value)
+      expect(TestModel.find_by(name: "One")).to eq :value
+
+      expect(SupportTableCache.cache).to receive(:fetch).and_return(:other_value)
+      expect(TestModel.find_by(code: "one", group: "First")).to eq :other_value
+
+      expect(SupportTableCache.cache).to_not receive(:fetch)
+      expect(TestModel.find_by(value: 1)).to eq record_1
+      expect(TestModel.find_by(name: "One", value: 1)).to eq record_1
+    end
+  end
+
+  describe "finding on a relation" do
     it "uses the cache when finding by multiple cacheable attributes with a relation chain" do
       expect(TestModel.where(group: "First").find_by(code: "one")).to eq record_1
       expect(SupportTableCache.cache.read(SupportTableCache.cache_key(TestModel, {code: "one", group: "First"}, ["code", "group"], false))).to eq record_1
@@ -72,16 +95,53 @@ describe SupportTableCache do
       expect(TestModel.where(code: "one").find_by(group: "First").value).to eq 1
     end
 
+    it "uses the cache when finding by multiple cacheable attributes with a relation chain with find_by!" do
+      expect(TestModel.where(group: "First").find_by!(code: "one")).to eq record_1
+      expect(SupportTableCache.cache.read(SupportTableCache.cache_key(TestModel, {code: "one", group: "First"}, ["code", "group"], false))).to eq record_1
+    end
+
+    it "raises an error when using find_by! on a relation and the record doesn't exist" do
+      expect { TestModel.where(group: "First").find_by!(code: "not exist") }.to raise_error(ActiveRecord::RecordNotFound)
+    end
+
     it "does not use the cache when finding by a non-cacheable attribute" do
-      expect(SupportTableCache.cache).to receive(:fetch).and_return(:value)
-      expect(TestModel.find_by(name: "One")).to eq :value
-
-      expect(SupportTableCache.cache).to receive(:fetch).and_return(:other_value)
-      expect(TestModel.find_by(code: "one", group: "First")).to eq :other_value
-
       expect(SupportTableCache.cache).to_not receive(:fetch)
-      expect(TestModel.find_by(value: 1)).to eq record_1
-      expect(TestModel.find_by(name: "One", value: 1)).to eq record_1
+      expect(TestModel.where(group: "First").find_by(value: 1)).to eq record_1
+    end
+  end
+
+  describe "fetching" do
+    it "fetches records by attributes" do
+      expect(TestModel.fetch_by(name: "One")).to eq record_1
+      expect(SupportTableCache.cache.read(SupportTableCache.cache_key(TestModel, {name: "One"}, ["name"], true))).to eq record_1
+    end
+
+    it "fetches scoped records by attributes" do
+      expect(TestModel.where(group: "First").fetch_by(code: "one")).to eq record_1
+      expect(SupportTableCache.cache.read(SupportTableCache.cache_key(TestModel, {code: "one", group: "First"}, ["code", "group"], false))).to eq record_1
+    end
+
+    it "fetches scoped records by attributes with an empty scope" do
+      expect(TestModel.readonly.fetch_by(name: "One")).to eq record_1
+      expect(SupportTableCache.cache.read(SupportTableCache.cache_key(TestModel, {name: "One"}, ["name"], true))).to eq record_1
+    end
+
+    it "raises an ArgumentError if fetched query is not cacheable" do
+      expect { TestModel.fetch_by(code: "one") }.to raise_error(ArgumentError)
+    end
+
+    it "raises an ArgumentError if scoped fetched query is not cacheable" do
+      expect { TestModel.where(value: nil).fetch_by(code: "one") }.to raise_error(ArgumentError)
+    end
+
+    it "raises an ActiveRecord::RecordNotFoundError if fetch_by! does not find a record" do
+      expect(TestModel.fetch_by!(name: "One")).to eq record_1
+      expect { TestModel.fetch_by!(name: "Not Exist") }.to raise_error(ActiveRecord::RecordNotFound)
+    end
+
+    it "raises an ActiveRecord::RecordNotFoundError if scoped fetch_by! does not find a record" do
+      expect(TestModel.where(group: "First").fetch_by!(code: "one")).to eq record_1
+      expect { TestModel.where(group: "First").fetch_by!(code: "Not Exist") }.to raise_error(ActiveRecord::RecordNotFound)
     end
   end
 
@@ -147,6 +207,20 @@ describe SupportTableCache do
       end
       expect(SupportTableCache.disabled?).to eq false
     end
+
+    it "can disable just for a class" do
+      SupportTableCache.disable do
+        TestModel.enable_caching do
+          expect(SupportTableCache.disabled?).to eq false
+          expect(SupportTableCache.cache).to receive(:fetch)
+          expect(TestModel.find_by(name: "One")).to eq record_1
+          TestModel.disable_caching do
+            expect(SupportTableCache.cache).not_to receive(:fetch)
+            expect(TestModel.find_by(name: "One")).to eq record_1
+          end
+        end
+      end
+    end
   end
 
   describe "setting the cache" do
@@ -159,6 +233,79 @@ describe SupportTableCache do
       ensure
         SupportTableCache.cache = cache
       end
+    end
+
+    it "can set a cache per class" do
+      cache = ActiveSupport::Cache::MemoryStore.new
+      TestModel.support_table_cache = cache
+      save_cache = SupportTableCache.cache
+      SupportTableCache.cache = nil
+      begin
+        expect(cache).to receive(:fetch).twice.and_call_original
+        expect(TestModel.find_by(name: "One")).to eq record_1
+        expect(TestModel.find_by(name: "One")).to eq record_1
+      ensure
+        SupportTableCache.cache = save_cache
+        TestModel.support_table_cache = nil
+      end
+    end
+
+    it "can set the cache to an in memory cache" do
+      save_cache = SupportTableCache.cache
+      begin
+        SupportTableCache.cache = :memory
+        expect(SupportTableCache.cache).to be_a(SupportTableCache::MemoryCache)
+        expect(SupportTableCache.cache.object_id).to_not eq save_cache.object_id
+      ensure
+        SupportTableCache.cache = save_cache
+      end
+    end
+
+    it "can set a cache per class to an in memory cache" do
+      TestModel.support_table_cache = :memory
+      expect(TestModel.send(:support_table_cache_impl)).to be_a(SupportTableCache::MemoryCache)
+    ensure
+      TestModel.support_table_cache = nil
+    end
+  end
+
+  describe "loading the cache" do
+    it "loads the cache with all records" do
+      cache = ActiveSupport::Cache::MemoryStore.new
+      TestModel.support_table_cache = cache
+      begin
+        TestModel.load_cache
+      ensure
+        TestModel.support_table_cache = nil
+      end
+
+      [
+        [record_1, {"name" => "One"}, ["name"], true],
+        [record_1, {"code" => "one", "group" => "First"}, ["code", "group"], false],
+        [record_2, {"name" => "Two"}, ["name"], true],
+        [record_2, {"code" => "two", "group" => "Second"}, ["code", "group"], false]
+      ].each do |record, attributes, attribute_names, case_sensitive|
+        cache_key = SupportTableCache.cache_key(TestModel, attributes, attribute_names, case_sensitive)
+        expect(cache.read(cache_key)).to eq record
+      end
+    end
+  end
+
+  describe "testing!" do
+    it "initializes new caches within a block" do
+      normal_cache = SupportTableCache.cache
+
+      SupportTableCache.testing! do
+        testing_cache = SupportTableCache.cache
+        expect(testing_cache).to be_a(SupportTableCache::MemoryCache)
+        expect(testing_cache).to_not eq normal_cache
+
+        SupportTableCache.testing! do
+          expect(SupportTableCache.cache).to eq testing_cache
+        end
+      end
+
+      expect(SupportTableCache.cache).to eq normal_cache
     end
   end
 end